ext/resources: early-stage JS messaging refactoring
starting to standardise the messaging format between the different JS
files, with the goal of being able to validate and type messages from
issuance to read in the Swift extension.

this commit also simplifies the view rendering logic and removes much of
the global state from background, content and popup, preferring a more
injected, localised and functional approach.